Add a way to provide global parameters to sema::FreeformGeometry
If I were designing this format, I would have, of course, specified
these parameters in a .geom file itself. But as I clearly wasn't doing
that, some parameters which are used to build `Struct` are not
encapsulated in the geometry file and are instead read from the main
configuration file in legacy code.