Do not store pointer to `FileView` in `parser::SourceRange`
These changes (1) shrink SourceRange's size; (2) make it easy to assign
special meaning to otherwise invalid combinations of (ptr, size) pair
stored inside it; (3) as a consequence of the previous point, make it
possible to not require file content to end with '\n'.

This is, unfortunately, quite massive commit, but it is a bit hard to do
this refactoring in a larger number of atomic commits (but, frankly,
nobody will ever look at the git history of this project).