Align cell shifts with intuitive behavior during Hamiltonian building
One would expect the affected fractional_shift_fp vector to point from
the 0th primitive cell to the cell where the hopping target is. This was
not the case previously, however: the vector actually pointed in the
opposite direction. While this didn't introduce any bugs (hopping are
symmetric wrt reflection), variable naming used in the function
afterwards suggested the opposite direction of fractional_shift_fp.